pyproject.toml
[dependency-groups]
dev = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "flask>=3.1.2",
    "magic-filter>=1.0.12",
    "pydantic>=2.11.10",
    "pydantic-core>=2.33.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "strict"


[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
  "E", "F",   # базовые ошибки/flake8
  "I",        # сортировка импортов
  "B",        # bugbear (частые баги)
  "UP",       # pyupgrade
]
ignore = [
  "E501",     # длина строки (лучше отдать форматтеру)
]

[tool.ruff.format]
quote-style = "double"
